+++ CHANGELOG.rst
@@ -4,9 +4,7 @@ changelog
 1.1 - XX May 2017
 ...........................................
 * fix race condition when capturing  queries not tokenised that leave the binglog position unchanged
-* fix race condition when capturing the queries with wrong binglog position
-* fix race condition when rotate event happens with the table_limit enabled
-
+* update docstrings in sql_util.py
 
 1.0 - 07 May 2017
 ............................................